==> Documents/1.manuscript/2.CC-RIX/2.Imputation_CC-RIX_old_for_code.Rmd <==
---
title: "local imputations"
author: Anna L Tyler
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output: 
  html_document:
    code_folding: hide
    collapsed: no
bibliography: hdm.bib 
---

```{r param}
rm(list = ls()) #clear out previous R environment
remove.pseudogenes = TRUE #if TRUE, removes pseudogenes and predicted genes 
tissue.names <- c("Adipose", "Islet", "Liver", "SkeletalMuscle")
tissue.cols <- c("orange", "#8dd3c7", "tan", "brown") #colors that are not related to the DO/CC colors
is.interactive = FALSE
overwrite.results = FALSE
```

## Introduction
In this workflow, we impute gene expression in the DO
using the effects of the local marker. We fit the measured
transcript abundance to the nearest marker with a kinship 
correction. We use the fitted values from that model as 
the imputed transcript values.

We also estimate measures of heritability, both overall and
distal. We calculate overall heritability using est_herit
on measured gene expression. We may not use this, but we 
want to have this. We calculate distal heritability by 
using est_herit on the residuals after fitting the linear 
model with the nearest marker. The product of the variance
of the residuals times their estimated heritability is the 
distal heritability component for the transcript.

We then compare these estimates of heritability to a measure 
of trait relevance, which is the maximum correlation of each
transcript with the traits.

We see that local heritability is anti-correlated with 
trait relevance. That is, transcripts whose abundance
is highly related to genotype at the local marker tend 
to have low correlations with traits. On the other hand,
transcripts with high trait relevance, tend to have low
local heritability.

We see a different relationship between trait relevance
and the distal component of heritability. Transcripts
with high trait relevance tend to have low-mid to mid-level 
distal heritability.

This workflow depends on results from 
1a.Tissue_Expression.Rmd, 1b.Trait_Selection.Rmd, and 
2a.Kinship_Expression_Traits.Rmd.

```{r here, message = FALSE, warning = FALSE}
library(here)
imputed.results.dir <- here("Results", "Imputed_Transcriptomes")
if(!file.exists(imputed.results.dir)){dir.create(imputed.results.dir, recursive = TRUE)}
data.results.dir <- here("Results", "CC-RIX", "Data") #a results folder that holds parsed data that we don't want to mix in with downloaded data
```

Load code and libraries.

```{r load_code, message = FALSE, warning = FALSE}
all.fun <- list.files(here("Code"), pattern = ".R", full.names = TRUE)
for(i in 1:length(all.fun)){source(all.fun[i])}

all.packages <- c("qtl2", "gprofiler2", "pheatmap")
load_libraries(all.packages)
```

Read in the adjusted CC-RIX transcriptomes generated by 1a.CC-RIX_Phenotypes_Expression.Rmd.
Mean center and scale the expression across individuals.

```{r read_transcriptomes}
#CC-RIX transcriptome files generated by 
#1a.CC-RIX_Phenotypes_Expression.Rmd
#these transcripts are normalized and adjusted
#for covariates, but are not mean centered or 
#scaled
adj.expr <- readRDS(file.path(data.results.dir, "Expression.by.Tissue.Adjusted.RDS")) 
tissue.names <- names(adj.expr)
tissue.cols <- c("orange", "tan", "brown") #colors that are not related to the DO/CC colors

#mean center and standardize gene expression across individuals so all variances are 1
scaled.expr <- lapply(adj.expr, function(x) apply(x, 2, scale))
for(tx in 1:length(scaled.expr)){
    dimnames(scaled.expr[[tx]]) <- dimnames(adj.expr[[tx]])
}

load(here("Data", "DO", "QTLViewer_Geno_V10.Rdata")) #ensembl.version, genoprobs, K, map, markers
covar <- readRDS(file.path(data.results.dir, "Clinical_Phenotype_Covariates.RDS")) #generated by 1b.Trait_Selection.Rmd
gene.tables <- readRDS(file.path(data.results.dir, "Gene_Tables.RDS")) #generated by 1a.Tissue_Expression.Rmd

#eventually use the coefficients generated by 3a.Imputation.Rmd
#for now we will use those in the QTLviewer objects
local.coef <- readRDS(here("Results", "DO", "Transcriptomes", "eQTL_coef_local.RDS")) #generated in 1a.Tissue_Expression.Rmd
```


## Impute gene expression

We impute gene expression in the CC-RIX individuals using the 
eQTL coefficients from the DO and the local genotype in the 
CC-RIX.


```{r impute_transcriptomes}
save.every = 100

for(tx in 1:length(tissue.names)){
    start.idx <- 1
    transcript.id <- colnames(adj.expr[[tx]])

    local.imp.file <- file.path(imputed.results.dir, paste0("Local_Imputation_Expr_", tissue.names[tx], ".RDS"))
    
    #check to see if we've already saved some results that we can use
    #we only need to use one of the files to figure this out
    if(file.exists(distal.herit.file) && !overwrite.results){
      herit.resid <- readRDS(distal.herit.file)
      start.idx <- max(which(!is.na(herit.resid[,1]))) + 1
      local.imp <- readRDS(local.imp.file)
    }
      
  #start where we left off, or at the beginning if there are no data
  #or we are overwriting previous data.
  if(start.idx < length(transcript.id)){
    sink(file.path(imputed.results.dir, paste0("progress_", tissue.names[tx], ".txt")))

    #if we are starting fresh...
    if(!file.exists(distal.herit.file) || overwrite.results){
      herit.resid <- matrix(NA, ncol = 3, nrow = length(transcript.id))
      rownames(herit.resid) <- transcript.id
      colnames(herit.resid) <- c("local_var", "residual_variance", "residual_heritability")

      local.imp <- matrix(NA, ncol = length(transcript.id), nrow = nrow(adj.expr[[tx]]))
      dimnames(local.imp) <- dimnames(adj.expr[[tx]])
    }
  
    #start where we left off, or at the beginning, if that is warranted
    for(tr in start.idx:length(transcript.id)){
      print(tr)
      
      gene.info.idx <- which(gene.tables[[tx]]$gene.id == transcript.id[tr])
      nearest.marker <- gene.tables[[tx]]$nearest.marker.id[gene.info.idx]

       #we won't be able to find nearest markers for genes on Y or MT chromosomes
      if(nearest.marker == ""){next()}  

      marker.geno <- pull_genoprobpos(genoprobs, marker = nearest.marker)

      trans.expr <- scaled.expr[[tx]][,transcript.id[tr]]
  
      #with kinship
      #I am keeping the variance of the fitted values to use as the local variance explained
      #reserve as much of the signal from kinship as we can by performing a kinship correction
      model.kin <- fit1(marker.geno, trans.expr, kinship = Kg)
      imp.local <- model.kin$fitted
      local.imp[,tr] <- imp.local
      distal.resid <- model.kin$resid
      result <- c(var(imp.local), var(distal.resid), est_herit(distal.resid, Kg))
      herit.resid[tr,] <- result

      if(tr %% save.every == 0){
        saveRDS(herit.resid, distal.herit.file)
        saveRDS(local.imp, local.imp.file)
      }
    }
    saveRDS(herit.resid, distal.herit.file)
    saveRDS(local.imp, local.imp.file)
    sink()
  }
}
```

```{r read_imp}
local.imp.files <- list.files(imputed.results.dir, pattern = "Local", full.names = TRUE)
tx.local.imp <- lapply(local.imp.files, readRDS)

distal.herit.files <- list.files(imputed.results.dir, pattern = "Distal", full.names = TRUE)
tx.distal.herit <- lapply(distal.herit.files, readRDS)

local.var.exp <- lapply(tx.distal.herit, function(x) x[,"local_var"])
distal.var.exp <- lapply(tx.distal.herit, function(x) x[,"residual_variance"]*x[,"residual_heritability"])
```


## Correlation Check

### Imputed and measured {.tabset .tabset-fade .tabset-pills}
 
Check the correlations of the imputed transcripts with
transcript measurements. Transcripts with higher local 
LOD scores should have higher correlations to the 
measured transcript.

```{r check_correlations, results = "asis", fig.height = 4, fig.width = 8}
for(tx in 1:length(tissue.names)){
    cat("####", tissue.names[tx], "\n")
    measured <- scaled.expr[[tx]]
    imputed <- tx.local.imp[[tx]]

    common.tx <- intersect(colnames(measured), colnames(imputed))
    common.ind <- intersect(rownames(measured), rownames(imputed))
    all.cor <- sapply(1:length(common.tx), 
        function(x) cor(measured[common.ind,common.tx[x]], 
        imputed[common.ind,common.tx[x]]))
    local.lod <- local.coef[[tx]]$lod[match(common.tx, local.coef[[tx]]$gene.id)]
    par(mfrow = c(1,2))
    hist(all.cor, main = "Histogram of Correlations", 
        xlab = "Correlation", xlim = c(0,1))
    plot(local.lod, all.cor, xlab = "local LOD score", 
        main = "Correlation v. LOD",
        ylab = "Correlation", ylim = c(0,1))
    mtext(tissue.names[tx], side = 3, outer = TRUE, line = -1.5, font = 2)
    cat("\n\n")

}
```

### Distal and local heritabilities

Check that local and distal heritabilities are anti-correlated
and never sum to more than 1.

```{r local_distal_check, fig.width = 8, fig.height = 8}

par(mfrow = c(2,2))
for(tx in 1:length(tissue.names)){
    #check.local.var <- apply(tx.local.imp[[tx]], 2, var)
    local.var <- tx.distal.herit[[tx]][,"local_var"]
    #plot(check.local.var, local.var)
    distal.var <- tx.distal.herit[[tx]][,"residual_variance"]
    plot(local.var, distal.var, xlab = "Local Variance Explained",
    ylab = "Distal Variance Explained", main = tissue.names[tx])
}

```

### Local and distal contribution to transcripts

What are the relative contributions of local and distal 
regulation to trait variation?

The following plot compares the variance explained by local and
distal components across the tissues. Both components contribute
about the same to transcript levels.

We calculated these values in the following way:

1. The variance of the measured transcript is the sum of the 
    variance contributions from covariates, local genotype, 
    distal genotype, and error. We already removed the variance
    from the covariates, so we are left with all the other 
    variance components. 

*transcript = cov + local_genotype + kinship + error*

2. We scaled the remaining variance each transcript (measured minus 
    covariates) to be equal to 1. Thus, the variance contributions
    from the local, distal, and error components sum to 1.

*adj_transcript = local_genotype + kinship + error*

3. We then fit this variance with a model using the local genotype
    and kinship matrix. The residual variance from this model is
    the variance explained by distal genetics and by error. We can
    estimate the proportion explained by distal genetics by estimating
    the heritability of this residual. The *distal variance explained* 
    is the variance of the residual of the above model times the 
    heritability of the residual. 
 

```{r local_var, fig.width = 8, fig.height = 5}

common.tx <- Reduce("intersect", lapply(distal.var.exp, names))
local.distal.comp <- list("Local" = Reduce("cbind", lapply(local.var.exp, function(x) x[common.tx])),
    "Distal" = Reduce("cbind", lapply(distal.var.exp, function(x) x[common.tx])))
for(i in 1:length(local.distal.comp)){
    colnames(local.distal.comp[[i]]) <- tissue.names
}
plot.grouped.boxes(local.distal.comp, type = "matrix", ylab = "Variance Explained",
    main = "Variance Explained by Local and Distal Components")
```


## Heritability and Trait Relevance

We want to examine the relationship between heritability and 
trait relevance. We define trait relevance here as the correlation
between the transcript and the trait. Heritability we break into
two components: local heritability and distal heritability.

Local heritability can be defined either with the local LOD
score, or the variance explained by the local marker. We will
use the variance explained by the local marker to be more analogous
both with the human literature and with the distal heritability
measure. Distal heritability is the variance eplained by the 
kinship matrix for the transcript levels after controlling for the
local marker.


### Local Variance Explained and Trait Relavance {.tabset .tabset-fade .tabset-pills}

The following plots show the correlation between the 
variance explained by the local marker and the trait
relevance (maximum correlation between the transcript
and all traits.)

#### Variance Explained - Loess

```{r local_v_traits, fig.height = 8, fig.width = 8}
trait.cor.files <- list.files(here("Results", "Transcriptomes"), 
    pattern = "Maximum_Trait_Correlation", full.names = TRUE) #generated in 2a.Kinship_Expression_Traits.Rmd
trait.cor <- lapply(trait.cor.files, readRDS)
names(trait.cor) <- tissue.names

par(mfrow = c(2,2))
for(tx in 1:length(trait.cor)){

    common.tx <- intersect(names(trait.cor[[tx]]), names(local.var.exp[[tx]])) #transcripts should be in the same order, but just make sure

    binned_ts <- bin.vector2(local.var.exp[[tx]][common.tx], seq(0, 1, 0.05))
    bin_means <- sapply(binned_ts, function(x) mean(abs(trait.cor[[tx]][names(x)]), na.rm = TRUE))
    x_means <- sapply(binned_ts, mean)

    plot(local.var.exp[[tx]][common.tx], abs(trait.cor[[tx]][common.tx]),
        xlab = "Local Variance Explained", ylab = "Max Correlation with Trait",
        main = tissue.names[tx], pch = 16)
    points(x_means, bin_means, col = "lightblue", type = "l", lwd = 3)
}
```

#### Variance Explained - Linear

The following plots are the same as above, but show the linear fit.

```{r local_lin_fit, fig.height = 8, fig.width = 8}
par(mfrow = c(2,2))
for(tx in 1:length(trait.cor)){
    common.tx <- intersect(names(trait.cor[[tx]]), names(local.var.exp[[tx]]))

    plot.with.model(local.var.exp[[tx]][common.tx], abs(trait.cor[[tx]][common.tx]),
        xlab = "Local Variance Explained", ylab = "Max Correlation with Trait",
        main = tissue.names[tx], report = "cor.test")
}
```

#### LOD - Linear

The following plots show the relationship between trait
relevance and local LOD score.

```{r herit_lod, fig.width = 8, fig.height = 8}
par(mfrow = c(2,2))
for(tx in 1:length(tissue.names)){
    local.lod <- unlist(local.coef[[tx]][,"lod"])
    names(local.lod) <- unlist(local.coef[[tx]][,"gene.id"])
    common.tx <- intersect(names(local.lod), names(trait.cor[[tx]]))

    plot.with.model(local.lod[common.tx], abs(trait.cor[[tx]][common.tx]),
        xlab = "Local Variance Explained", ylab = "Max Correlation with Trait",
        main = tissue.names[tx], report = "cor.test")
}

```

### Distal Variance Explained and Trait Relavance {.tabset .tabset-fade .tabset-pills}

#### Variance Explained - Loess

```{r, fig.width = 8, fig.height = 8}

par(mfrow = c(2,2))
for(tx in 1:length(trait.cor)){
    distal.herit <- tx.distal.herit[[tx]][,"residual_variance"]*tx.distal.herit[[tx]][,"residual_heritability"]
    common.tx <- intersect(names(trait.cor[[tx]]), names(distal.herit))

    binned_ts <- bin.vector2(distal.herit[common.tx], seq(0, 1, 0.05))
    bin_means <- sapply(binned_ts, function(x) mean(abs(trait.cor[[tx]][names(x)]), na.rm = TRUE))
    x_means <- sapply(binned_ts, mean)

    plot(distal.herit[common.tx], abs(trait.cor[[tx]][common.tx]),
        xlab = "Variance Explained", ylab = "Max Correlation with Trait",
        main = tissue.names[tx], pch = 16)
    points(x_means, bin_means, col = "lightblue", type = "l", lwd = 3)
}
```

#### Variance Explained - Linear

The following figures show the same data, but with linear fit models.

```{r lin_fig, fig.width = 8, fig.height = 8}
par(mfrow = c(2,2))
for(tx in 1:length(trait.cor)){
    distal.herit <- tx.distal.herit[[tx]][,"residual_variance"]*tx.distal.herit[[tx]][,"residual_heritability"]
    common.tx <- intersect(names(trait.cor[[tx]]), names(distal.herit))

    plot.with.model(distal.herit[common.tx], abs(trait.cor[[tx]][common.tx]),
        xlab = "Variance Explained", ylab = "Max Correlation with Trait",
        main = tissue.names[tx], report = "cor.test")
}
```

## Distal Tails {.tabset .tabset-fade .tabset-pills}

The plots with distal variance explained vs. trait relevance
have prominant tails with high variance explained but very low
trait relevance. What are these genes?

```{r tails, fig.width = 10, fig.height = 4.5, results = "asis"}

tail.boundaries <- matrix(c(0.5, 0.3, 0.7, 0.2, 0.5, 0.3, 0.55, 0.2), ncol = 2, byrow = TRUE)
rownames(tail.boundaries) <- tissue.names
colnames(tail.boundaries) <- c("Variance Explained", "Trait Relatedness")

tail.enrich <- tail.genes <- vector(mode = "list", length = length(tissue.names))
names(tail.enrich) <- names(tail.genes) <- tissue.names

#pdf("~/Desktop/tails.pdf", width = 8, height = 5)
for(tx in 1:length(trait.cor)){
    cat("###", tissue.names[tx], "\n")

    distal.herit <- tx.distal.herit[[tx]][,"residual_variance"]*tx.distal.herit[[tx]][,"residual_heritability"]
    common.tx <- intersect(names(trait.cor[[tx]]), names(distal.herit))

    tail.idx <- intersect(which(distal.herit[common.tx] > tail.boundaries[tx,1]), 
        which(abs(trait.cor[[tx]][common.tx]) < tail.boundaries[tx,2]))
    tail.genes[[tx]] <- common.tx[tail.idx]
    tail.enrich[[tx]] <- gost(tail.genes[[tx]], organism = "mmusculus", 
        sources = c("GO", "REACTOME", "HP", "CORuM"))


    pt.col <- rep("black", length(common.tx))
    pt.col[tail.idx] <- "salmon"

    layout(matrix(c(1,2), ncol = 2), widths = c(0.8, 1))

    par(mar = c(4,4,4,4))
    plot.with.model(distal.herit[common.tx], abs(trait.cor[[tx]][common.tx]),
        xlab = "Variance Explained", ylab = "Max Correlation with Trait",
        main = tissue.names[tx], report = "cor.test", col = pt.col)

    plot.enrichment(tail.enrich[[tx]], order.by = "p_value", num.terms = 20, max.term.size = 3000,
        plot.label = paste("Tail Enrichment for", tissue.names[tx]), text.size = 0.7)

    cat("\n\n")
}
#dev.off()
```

### All Tissues

All the tails are enriched for ribosomal genes.

```{r all_enrich, fig.height = 7, fig.width = 5}
plot.enrichment.group(tail.enrich, plot.label = "All Tail Enrichments", max.term.size = 3000,
    n.terms = 20)
```

### All Genes

The individual gene names in each tail are shown below

```{r tail_genes, fig.width = 3, fig.height = 15}
all.tail <- Reduce("union", tail.genes)

tail.table <- matrix(0, nrow = length(all.tail), ncol = length(tissue.names))
rownames(tail.table) <- all.tail
colnames(tail.table) <- tissue.names

for(tx in 1:length(tissue.names)){
    tail.table[tail.genes[[tx]],tx] <- 1
}

gene.info  <- cbind(Reduce("c", lapply(gene.tables, function(x) x$symbol)), 
    Reduce("c", lapply(gene.tables, function(x) x$gene.id)))
gene.names  <- gene.info[match(all.tail, gene.info[,2]),1]
rownames(tail.table) <- gene.names

#pdf("~/Desktop/tail.genes.pdf", width = 3, height = 15)
pheatmap(tail.table, cluster_cols = FALSE, cex = 0.7)
#dev.off()


#gene.order <- order(rownames(tail.table))
#pheatmap(tail.table[gene.order,], cluster_rows = FALSE, cluster_cols = FALSE, cex = 0.7)

#par(mfrow = c(2,2))
#for(tx in 1:length(tissue.names)){
#    var.test <- apply(adj.expr[[tx]], 2, var)
#    not.tail <- setdiff(colnames(adj.expr[[tx]]), tail.genes[[tx]])
#    boxplot(list("Tail Genes" = log10(var.test[tail.genes[[tx]]]), 
#        "Other Genes" = log10(var.test[not.tail])), main = tissue.names[tx])
#}
```
